fix #11 and fix #13
Vorple:
Removed the line that said you can retrieve the truth state returned by the JS command since you actually can't. Removed a superfluous period when showing the date with a non-Vorple interpreter in "The Grandfather Clock" example.

Vorple Screen Effects:
Added the word "font" in the documentation when mentionning the style, since they have that word in their name.